tests/Chat/requirements.txt
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
cassandra-driver>=3.28.0
PyYAML>=6.0
httpx>=0.24.0

